requirements.txt
termcolor==1.1.0
PyYAML==3.12
